Made function definition of intersection and reach lib functions either inline or moved them to cpp file.
